# my-blog-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  my-blog-backend_internal_api_v1_dto_request.CreateCommentRequest:
    properties:
      articleId:
        type: integer
      content:
        maxLength: 1000
        minLength: 1
        type: string
      parentId:
        type: integer
    required:
    - articleId
    - content
    type: object
...
  my-blog-backend_internal_api_v1_dto_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@myblog.com
    name: API Support
    url: http://www.myblog.com/support
  description: This is a blog backend API with Swagger documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: My Blog API
  version: "1.0"
paths:
  /admin/comments:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 获取所有评论列表（管理员）
      tags:
      - 评论管理
  /admin/comments/{id}/approve:
    put:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 审核通过评论（管理员）
      tags:
      - 评论管理
  /admin/comments/{id}/reject:
    put:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 审核拒绝评论（管理员）
      tags:
      - 评论管理
  /admin/comments/pending:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 获取待审核评论列表（管理员）
      tags:
      - 评论管理
...
  /articles/{id}/comments:
    get:
      consumes:
      - application/json
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取文章评论列表
      tags:
      - 评论管理
  /comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - 评论管理
  /comments/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取评论详情
      tags:
      - 评论管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_request.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 更新评论
      tags:
      - 评论管理
  /comments/{id}/like:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 点赞评论
      tags:
      - 评论管理
  /comments/{id}/replies:
    get:
      consumes:
      - application/json
      parameters:
      - description: 父评论ID
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取子评论列表
      tags:
      - 评论管理
  /comments/{id}/replies/count:
    get:
      consumes:
      - application/json
      parameters:
      - description: 父评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取子评论数量
      tags:
      - 评论管理
  /comments/{id}/unlike:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      security:
      - BearerAuth: []
      summary: 取消点赞评论
      tags:
      - 评论管理
  /users/{user_id}/comments:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/my-blog-backend_internal_api_v1_dto_response.Response'
      summary: 获取用户评论列表
      tags:
      - 评论管理
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"